ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
transformers==4.35.0
pytorch-crf==0.7.2
sentencepiece==0.1.99
numpy==1.24.3
pandas==2.0.3
gunicorn==21.2.0